sprinkle lighthouse into the browserspec tests
