+++ src/app/content/content.browserspec.ts
@@ -1,5 +1,5 @@
 /** @jest-environment puppeteer */
-import { finishRender, navigate } from '../../test/browserutils';
+import { finishRender, navigate, checkLighthouse } from '../../test/browserutils';
 
 const TEST_PAGE = '/books/testbook1-shortid/pages/testpage1-shortid';
 
@@ -33,4 +33,9 @@ describe('content', () => {
     });
     expect(isSkipToContentSelected).toBe(true);
   });
+
+  it('a11y lighthouse check', async() => {
+    await checkLighthouse(TEST_PAGE);
+  });
+
 });
